ResumeAnalyzer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume-analyzer"
version = "1.0.0"
description = "AI-powered resume search and analysis using RAG (Retrieval-Augmented Generation)"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Resume Analyzer Team"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "python-multipart>=0.0.22",
    "pydantic>=2.4.2",
    
    # LangChain & AI
    "langchain>=1.2.12",
    "langchain-core>=1.2.18",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    
    # LLM Providers
    "langchain-mistralai>=1.1.1",
    "langchain-openai>=1.1.11",
    "langchain-groq>=1.1.2",
    "langchain-openrouter>=0.1.0",
    
    # Database & Vector Store
    "langchain-mongodb>=0.11.0",
    "pymongo>=4.16.0",
    
    # Document Processing
    "pypdf>=6.8.0",
    
    # Frontend
    "streamlit>=1.55.0",
    "requests>=2.31.0",
    
    # Utilities
    "python-dotenv>=1.2.2",
    "tenacity>=8.2.3",
    "httpx>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/resume-analyzer"
Documentation = "https://github.com/yourusername/resume-analyzer#readme"

[tool.uv]
python-version = "3.14"

[tool.black]
line-length = 100
target-version = ['py312', 'py313', 'py314']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
